requirements.txt
pytrends
requests
google-genai
Pillow